+++ docs/nitpick-exceptions.ini
@@ -1,11 +1,11 @@
 [default]
 class_references=
     ; TODO: Understand why sphinx is not able to find this local class
-    opentelemetry.trace.propagation.textmap.TextMapPropagator
+    opentelemetry.propagators.textmap.TextMapPropagator
     ; - AwsXRayFormat
-    opentelemetry.trace.propagation.textmap.DictGetter
+    opentelemetry.propagators.textmap.DictGetter
     ; API
-    opentelemetry.trace.propagation.textmap.Getter
+    opentelemetry.propagators.textmap.Getter
     ; - DatadogFormat
     ; - AWSXRayFormat
     opentelemetry.sdk.trace.id_generator.IdGenerator
@@ -15,7 +15,7 @@ class_references=
 
 anys=
     ; API
-    opentelemetry.trace.propagation.textmap.TextMapPropagator.fields
+    opentelemetry.propagators.textmap.TextMapPropagator.fields
     ; - AWSXRayFormat
     TraceId
     ; - AwsXRayIdGenerator
