+++ propagator/opentelemetry-propagator-ot-trace/src/opentelemetry/propagators/ot_trace/__init__.py
@@ -17,6 +17,12 @@ from typing import Iterable, Optional
 
 from opentelemetry.baggage import get_all, set_baggage
 from opentelemetry.context import Context
+from opentelemetry.propagators.textmap import (
+    Getter,
+    Setter,
+    TextMapPropagator,
+    TextMapPropagatorT,
+)
 from opentelemetry.trace import (
     INVALID_SPAN_ID,
     INVALID_TRACE_ID,
@@ -25,12 +31,6 @@ from opentelemetry.trace import (
     TraceFlags,
     get_current_span,
     set_span_in_context,
-)
-from opentelemetry.trace.propagation.textmap import (
-    Getter,
-    Setter,
-    TextMapPropagator,
-    TextMapPropagatorT,
 )
 
 OT_TRACE_ID_HEADER = "ot-tracer-traceid"
@@ -153,7 +153,7 @@ class OTTracePropagator(TextMapPropagator):
         """Returns a set with the fields set in `inject`.
 
         See
-        `opentelemetry.trace.propagation.textmap.TextMapPropagator.fields`
+        `opentelemetry.propagators.textmap.TextMapPropagator.fields`
         """
         return {
             OT_TRACE_ID_HEADER,
